+++ packages/core/src/index.ts
@@ -32,3 +32,4 @@ export * from "./revision-requests.js";
 export * from "./handover.js";
 export * from "./discovery-template.js";
 export * from "./discovery.js";
+export * from "./mfa-policy.js";